% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ath_plantcyc.database0.0.1}
\alias{ath_plantcyc.database0.0.1}
\title{ath plantcyc database 0.0.1}
\format{
A databaseClass object.
}
\usage{
ath_plantcyc.database0.0.1
}
\description{
ath plantcyc database 0.0.1
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
