% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class.database}
\alias{class.database}
\title{pm Database classyFire 0.0.1}
\format{
A databaseClass object.
}
\usage{
class.database
}
\description{
in-build pmDatabse classyFire
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
