% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class.database_corncyc}
\alias{class.database_corncyc}
\title{class.database_corncyc}
\format{
A dataframe.
}
\usage{
class.database_corncyc
}
\description{
class database for corncyc
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
