% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_plant_database0.0.1}
\alias{kegg_plant_database0.0.1}
\title{Plant kegg database 0.0.1}
\format{
A databaseClass object.
}
\usage{
kegg_plant_database0.0.1
}
\description{
kegg_plant_database0.0.1
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
