% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_zma_pathway}
\alias{kegg_zma_pathway}
\title{kegg_zma_pathway}
\format{
A pathway_database object.
}
\usage{
kegg_zma_pathway
}
\description{
kegg_zma_pathway
}
\author{
Xiao Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
