% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_agri_database0.0.1}
\alias{knapsack_agri_database0.0.1}
\title{knapsack agri database 0.0.1}
\format{
A databaseClass object.
}
\usage{
knapsack_agri_database0.0.1
}
\description{
knapsack agri database 0.0.1
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
