% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{massbank_database0.0.4}
\alias{massbank_database0.0.4}
\title{massbank database database 0.0.4}
\format{
A databaseClass object.
}
\usage{
massbank_database0.0.4
}
\description{
massbank database database 0.0.4
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
