% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabobase_database0.0.1}
\alias{metabobase_database0.0.1}
\title{metaboBASE database 1.0.0}
\format{
A databaseClass object.
}
\usage{
metabobase_database0.0.1
}
\description{
metaboBASE database 1.0.0
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
