% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mona_database0.0.4}
\alias{mona_database0.0.4}
\title{MoNA - MassBank of North America database 0.0.4}
\format{
A databaseClass object.
}
\usage{
mona_database0.0.4
}
\description{
MassBank of North America database with classification
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
