% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plasma_database0.0.1}
\alias{plasma_database0.0.1}
\title{PlaSMA_database0.0.1}
\format{
A databaseClass object.
}
\usage{
plasma_database0.0.1
}
\description{
PlaSMA_database0.0.1
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
