% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{respect_database0.0.1}
\alias{respect_database0.0.1}
\title{respect_database0.0.1}
\format{
A databaseClass object.
}
\usage{
respect_database0.0.1
}
\description{
respect_database0.0.1
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
