% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zma_ms1_db_v0.0.1}
\alias{zma_ms1_db_v0.0.1}
\title{zma_ms1_db_v0.0.1}
\format{
A databaseClass object.
}
\usage{
zma_ms1_db_v0.0.1
}
\description{
zma_ms1_db_v0.0.1
}
\author{
Xiao Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
