% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zma_plantcyc_database0.0.1}
\alias{zma_plantcyc_database0.0.1}
\title{zea mays plantcyc database 0.0.1}
\format{
A databaseClass object.
}
\usage{
zma_plantcyc_database0.0.1
}
\description{
zea mays plantcyc database 0.0.1
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
\keyword{datasets}
